test(web-ui): cover api.ts request layer + consolidate split Jest roots (#774)
Frontend test-trust hardening (P3.3):

1. api.ts contract test — src/__tests__/lib/api.contract.test.ts pins
   method+URL+params+body for the critical endpoints (workspace/tasks/events
   + the auth request interceptor) via an axios-stub adapter, so wrong-endpoint
   / renamed-param / reshaped-body drift fails here instead of passing green
   under the module mocks every other suite uses.

2. Orphaned e2e suite removed — web-ui/e2e (no CI job, fully mocked backend,
   drifted fixtures) and its web-ui/playwright.config.ts are deleted. The
   real-backend Playwright harness at tests/e2e/ (CI-wired: e2e-browser-smoke
   + e2e-browser-full, #684/#703) already covers these pages. Dropped the dead
   test:e2e* scripts and the now-transitive @playwright/test devDep.

3. Jest roots consolidated under src/__tests__ — the legacy web-ui/__tests__
   root is gone. The 7 "duplicated" files were divergent, not redundant: 6
   pairs test disjoint concerns (e.g. PRDHeader header-actions vs stress-test
   button; TaskCard full-card vs status-tooltip; PRStatusPanel merge-gate vs
   CI/SWR/review). Kept both copies under focus-suffixed names
   (X.navigation/.actions/.events/.checks/.tooltip.test), matching the repo's
   existing multi-file-per-component convention — no coverage dropped. Only
   AssociatedTasksSummary (src ⊇ root) collapsed to a single file.

Suite: 95 → 95 suites, 1052 → 1065 tests, all green. Lint + build pass.

Closes #774
