+++ web-ui/jest.config.js
@@ -18,8 +18,7 @@ const customJestConfig = {
   testPathIgnorePatterns: [
     '/node_modules/',
     '/.next/',
-    '/e2e/',                          // Playwright tests — run via test:e2e, not Jest
-    '__tests__/utils/test-helpers',   // Shared test utilities, not a test suite
+    'src/__tests__/utils/test-helpers',  // Shared test utilities, not a test suite
   ],
   collectCoverageFrom: [
     'src/**/*.{js,jsx,ts,tsx}',
