+++ web-ui/package.json
@@ -9,10 +9,7 @@
     "lint": "eslint . --max-warnings 0",
     "test": "jest",
     "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage",
-    "test:e2e": "npx playwright test",
-    "test:e2e:ui": "npx playwright test --ui",
-    "test:e2e:headed": "npx playwright test --headed"
+    "test:coverage": "jest --coverage"
   },
   "dependencies": {
     "@hugeicons/core-free-icons": "^4.2.2",
@@ -44,7 +41,6 @@
   "devDependencies": {
     "@eslint/eslintrc": "^3.3.6",
     "@next/eslint-plugin-next": "^16.2.10",
-    "@playwright/test": "^1.61.1",
     "@tailwindcss/postcss": "^4.3.2",
     "@testing-library/jest-dom": "^6.4.0",
     "@testing-library/react": "^16.3.2",
