+++ web-ui/src/__tests__/components/tasks/TaskCard.test.tsx
@@ -1,96 +1,294 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import { TaskCard } from '@/components/tasks/TaskCard';
-import { STATUS_INFO } from '@/lib/taskStatusInfo';
-import type { Task } from '@/types';
+import type { Task, TaskCostEntry } from '@/types';
 
-jest.mock('next/link', () => {
-  const MockLink = ({ href, children }: { href: string; children: React.ReactNode }) => (
-    <a href={href}>{children}</a>
-  );
-  MockLink.displayName = 'MockLink';
-  return MockLink;
-});
+// ─── Fixtures ───────────────────────────────────────────────────────
 
-// Radix UI tooltips use portals and pointer events that don't work in jsdom.
-// Replace with a simple always-visible version to test content.
-jest.mock('@/components/ui/tooltip', () => ({
-  TooltipProvider: ({ children }: { children: React.ReactNode }) => <>{children}</>,
-  Tooltip: ({ children }: { children: React.ReactNode }) => <>{children}</>,
-  TooltipTrigger: ({ children }: { children: React.ReactNode }) => <>{children}</>,
-  TooltipContent: ({ children }: { children: React.ReactNode }) => (
-    <div role="tooltip">{children}</div>
-  ),
-}));
+function makeTask(overrides: Partial<Task> = {}): Task {
+  return {
+    id: 'task-1',
+    title: 'Implement login',
+    description: 'Build the user login form with validation.',
+    status: 'READY',
+    priority: 1,
+    depends_on: [],
+    estimated_hours: 4,
+    created_at: '2026-01-01T00:00:00Z',
+    updated_at: '2026-01-02T00:00:00Z',
+    ...overrides,
+  };
+}
 
-const baseTask: Task = {
-  id: 'task-1',
-  title: 'Test Task',
-  description: 'A test task description',
-  status: 'BACKLOG',
-  priority: 0,
-  depends_on: [],
-};
-
-const defaultProps = {
-  task: baseTask,
-  selectionMode: false,
-  selected: false,
+const defaultHandlers = {
   onToggleSelect: jest.fn(),
   onClick: jest.fn(),
   onExecute: jest.fn(),
   onMarkReady: jest.fn(),
+  onStop: jest.fn(),
+  onReset: jest.fn(),
 };
 
-describe('TaskCard status badge tooltip', () => {
-  it('renders the status badge label', () => {
-    render(<TaskCard {...defaultProps} />);
-    expect(screen.getByText('Backlog')).toBeInTheDocument();
+function renderCard(taskOverrides: Partial<Task> = {}, props: Partial<Parameters<typeof TaskCard>[0]> = {}) {
+  const task = makeTask(taskOverrides);
+  return render(
+    <TaskCard
+      task={task}
+      selectionMode={false}
+      selected={false}
+      {...defaultHandlers}
+      {...props}
+    />
+  );
+}
+
+beforeEach(() => {
+  jest.clearAllMocks();
+});
+
+// ─── Tests ──────────────────────────────────────────────────────────
+
+describe('TaskCard', () => {
+  it('renders task title and description', () => {
+    renderCard();
+    expect(screen.getByText('Implement login')).toBeInTheDocument();
+    expect(screen.getByText('Build the user login form with validation.')).toBeInTheDocument();
   });
 
-  it('renders tooltip with BACKLOG meaning', () => {
-    render(<TaskCard {...defaultProps} />);
-    expect(screen.getByRole('tooltip')).toHaveTextContent(STATUS_INFO.BACKLOG.meaning);
+  it('renders status badge with correct label', () => {
+    renderCard({ status: 'IN_PROGRESS' });
+    expect(screen.getByText('In Progress')).toBeInTheDocument();
   });
 
-  it('renders tooltip with BACKLOG next steps', () => {
-    render(<TaskCard {...defaultProps} />);
-    expect(screen.getByRole('tooltip')).toHaveTextContent(STATUS_INFO.BACKLOG.nextSteps);
+  it('shows Execute button for READY tasks', () => {
+    renderCard({ status: 'READY' });
+    expect(screen.getByRole('button', { name: /execute/i })).toBeInTheDocument();
+    expect(screen.queryByRole('button', { name: /mark ready/i })).not.toBeInTheDocument();
   });
 
-  it('renders tooltip with READY meaning', () => {
-    const task = { ...baseTask, status: 'READY' as const };
-    render(<TaskCard {...defaultProps} task={task} />);
-    expect(screen.getByRole('tooltip')).toHaveTextContent(STATUS_INFO.READY.meaning);
+  it('shows Mark Ready button for BACKLOG tasks', () => {
+    renderCard({ status: 'BACKLOG' });
+    expect(screen.getByRole('button', { name: /mark ready/i })).toBeInTheDocument();
+    expect(screen.queryByRole('button', { name: /execute/i })).not.toBeInTheDocument();
   });
 
-  it('renders tooltip with FAILED meaning', () => {
-    const task = { ...baseTask, status: 'FAILED' as const };
-    render(<TaskCard {...defaultProps} task={task} onReset={jest.fn()} />);
-    expect(screen.getByRole('tooltip')).toHaveTextContent(STATUS_INFO.FAILED.meaning);
+  it('hides action buttons for non-actionable statuses', () => {
+    renderCard({ status: 'DONE' });
+    expect(screen.queryByRole('button', { name: /execute/i })).not.toBeInTheDocument();
+    expect(screen.queryByRole('button', { name: /mark ready/i })).not.toBeInTheDocument();
   });
 
-  it('renders tooltip with IN_PROGRESS meaning', () => {
-    const task = { ...baseTask, status: 'IN_PROGRESS' as const };
-    render(<TaskCard {...defaultProps} task={task} onStop={jest.fn()} />);
-    expect(screen.getByRole('tooltip')).toHaveTextContent(STATUS_INFO.IN_PROGRESS.meaning);
+  it('shows dependency indicator when task has dependencies', () => {
+    renderCard({ depends_on: ['dep-1', 'dep-2'] });
+    // Dependency count badge (Radix Tooltip — title attr removed, count is always visible)
+    expect(screen.getByText('2')).toBeInTheDocument();
   });
 
-  it('renders tooltip with DONE meaning', () => {
-    const task = { ...baseTask, status: 'DONE' as const };
-    render(<TaskCard {...defaultProps} task={task} />);
-    expect(screen.getByRole('tooltip')).toHaveTextContent(STATUS_INFO.DONE.meaning);
+  it('hides dependency indicator when task has no dependencies', () => {
+    renderCard({ depends_on: [] });
+    expect(screen.queryByTitle(/depends on/i)).not.toBeInTheDocument();
   });
 
-  it('renders tooltip with BLOCKED meaning', () => {
-    const task = { ...baseTask, status: 'BLOCKED' as const };
-    render(<TaskCard {...defaultProps} task={task} />);
-    expect(screen.getByRole('tooltip')).toHaveTextContent(STATUS_INFO.BLOCKED.meaning);
+  it('shows checkbox in selection mode', () => {
+    renderCard({}, { selectionMode: true });
+    expect(screen.getByRole('checkbox', { name: /select implement login/i })).toBeInTheDocument();
   });
 
-  it('renders tooltip with MERGED meaning', () => {
-    const task = { ...baseTask, status: 'MERGED' as const };
-    render(<TaskCard {...defaultProps} task={task} />);
-    expect(screen.getByRole('tooltip')).toHaveTextContent(STATUS_INFO.MERGED.meaning);
+  it('hides checkbox when not in selection mode', () => {
+    renderCard({}, { selectionMode: false });
+    expect(screen.queryByRole('checkbox')).not.toBeInTheDocument();
+  });
+
+  it('calls onClick when card is clicked', async () => {
+    const user = userEvent.setup();
+    renderCard();
+    await user.click(screen.getByText('Implement login'));
+    expect(defaultHandlers.onClick).toHaveBeenCalledWith('task-1');
+  });
+
+  it('calls onExecute without triggering onClick', async () => {
+    const user = userEvent.setup();
+    renderCard({ status: 'READY' });
+    await user.click(screen.getByRole('button', { name: /execute/i }));
+    expect(defaultHandlers.onExecute).toHaveBeenCalledWith('task-1');
+    expect(defaultHandlers.onClick).not.toHaveBeenCalled();
+  });
+
+  it('calls onMarkReady without triggering onClick', async () => {
+    const user = userEvent.setup();
+    renderCard({ status: 'BACKLOG' });
+    await user.click(screen.getByRole('button', { name: /mark ready/i }));
+    expect(defaultHandlers.onMarkReady).toHaveBeenCalledWith('task-1');
+    expect(defaultHandlers.onClick).not.toHaveBeenCalled();
+  });
+
+  it('has accessible role, tabindex, and aria-label', () => {
+    renderCard();
+    const card = screen.getByRole('button', { name: /view details for implement login/i });
+    expect(card).toBeInTheDocument();
+    expect(card).toHaveAttribute('tabindex', '0');
+  });
+
+  it('triggers onClick on Enter key press', async () => {
+    const user = userEvent.setup();
+    renderCard();
+    const card = screen.getByRole('button', { name: /view details for implement login/i });
+    card.focus();
+    await user.keyboard('{Enter}');
+    expect(defaultHandlers.onClick).toHaveBeenCalledWith('task-1');
+  });
+
+  it('triggers onClick on Space key press', async () => {
+    const user = userEvent.setup();
+    renderCard();
+    const card = screen.getByRole('button', { name: /view details for implement login/i });
+    card.focus();
+    await user.keyboard(' ');
+    expect(defaultHandlers.onClick).toHaveBeenCalledWith('task-1');
+  });
+
+  it('renders all status variants correctly', () => {
+    const statuses = ['BACKLOG', 'READY', 'IN_PROGRESS', 'DONE', 'BLOCKED', 'FAILED'] as const;
+    const labels = ['Backlog', 'Ready', 'In Progress', 'Done', 'Blocked', 'Failed'];
+
+    statuses.forEach((status, i) => {
+      const { unmount } = renderCard({ status });
+      expect(screen.getByText(labels[i])).toBeInTheDocument();
+      unmount();
+    });
+  });
+
+  // ─── Stop / Reset action tests ──────────────────────────────────────
+
+  it('shows Stop button for IN_PROGRESS tasks', () => {
+    renderCard({ status: 'IN_PROGRESS' });
+    expect(screen.getByRole('button', { name: /stop/i })).toBeInTheDocument();
+  });
+
+  it('does not show Stop button for non-IN_PROGRESS tasks', () => {
+    renderCard({ status: 'READY' });
+    expect(screen.queryByRole('button', { name: /stop/i })).not.toBeInTheDocument();
+  });
+
+  it('calls onStop without triggering onClick', async () => {
+    const user = userEvent.setup();
+    renderCard({ status: 'IN_PROGRESS' });
+    await user.click(screen.getByRole('button', { name: /stop/i }));
+    expect(defaultHandlers.onStop).toHaveBeenCalledWith('task-1');
+    expect(defaultHandlers.onClick).not.toHaveBeenCalled();
+  });
+
+  it('shows Reset button for FAILED tasks', () => {
+    renderCard({ status: 'FAILED' });
+    expect(screen.getByRole('button', { name: /reset/i })).toBeInTheDocument();
+  });
+
+  it('does not show Reset button for non-FAILED tasks', () => {
+    renderCard({ status: 'READY' });
+    expect(screen.queryByRole('button', { name: /reset/i })).not.toBeInTheDocument();
+  });
+
+  it('calls onReset without triggering onClick', async () => {
+    const user = userEvent.setup();
+    renderCard({ status: 'FAILED' });
+    await user.click(screen.getByRole('button', { name: /reset/i }));
+    expect(defaultHandlers.onReset).toHaveBeenCalledWith('task-1');
+    expect(defaultHandlers.onClick).not.toHaveBeenCalled();
+  });
+
+  it('Stop button has destructive ghost styling', () => {
+    renderCard({ status: 'IN_PROGRESS' });
+    const stopBtn = screen.getByRole('button', { name: /stop/i });
+    expect(stopBtn).toHaveClass('text-destructive');
+  });
+
+  // ─── Loading state tests ──────────────────────────────────────────
+
+  it('shows loading spinner instead of action button when isLoading', () => {
+    renderCard({ status: 'IN_PROGRESS' }, { isLoading: true });
+    // Should show spinner, not the Stop button
+    expect(screen.queryByRole('button', { name: /stop/i })).not.toBeInTheDocument();
+    expect(screen.getByRole('status', { name: /loading/i })).toBeInTheDocument();
+  });
+
+  it('shows action buttons when isLoading is false', () => {
+    renderCard({ status: 'IN_PROGRESS' }, { isLoading: false });
+    expect(screen.getByRole('button', { name: /stop/i })).toBeInTheDocument();
+  });
+
+  it('shows loading spinner for READY task when isLoading', () => {
+    renderCard({ status: 'READY' }, { isLoading: true });
+    expect(screen.queryByRole('button', { name: /execute/i })).not.toBeInTheDocument();
+    expect(screen.getByRole('status', { name: /loading/i })).toBeInTheDocument();
+  });
+
+  it('shows loading spinner for FAILED task when isLoading', () => {
+    renderCard({ status: 'FAILED' }, { isLoading: true });
+    expect(screen.queryByRole('button', { name: /reset/i })).not.toBeInTheDocument();
+    expect(screen.getByRole('status', { name: /loading/i })).toBeInTheDocument();
+  });
+
+  // ─── Cost badge (issue #558) ─────────────────────────────────────────
+
+  function makeCostMap(entries: Partial<TaskCostEntry>[]): Map<string, TaskCostEntry> {
+    const map = new Map<string, TaskCostEntry>();
+    for (const entry of entries) {
+      const full: TaskCostEntry = {
+        task_id: entry.task_id ?? 'task-1',
+        task_title: entry.task_title ?? 'Implement login',
+        agent_id: entry.agent_id ?? 'react-agent',
+        input_tokens: entry.input_tokens ?? 1000,
+        output_tokens: entry.output_tokens ?? 500,
+        total_cost_usd: entry.total_cost_usd ?? 0.12,
+      };
+      map.set(full.task_id, full);
+    }
+    return map;
+  }
+
+  it('renders a cost badge when cost data exists for the task', () => {
+    renderCard({}, {
+      costMap: makeCostMap([{ task_id: 'task-1', total_cost_usd: 0.1234 }]),
+    });
+    const badge = screen.getByTestId('cost-badge');
+    expect(badge).toBeInTheDocument();
+    expect(badge.textContent).toContain('$0.12');
+  });
+
+  it('hides the cost badge when no cost data exists for the task', () => {
+    renderCard({}, {
+      costMap: makeCostMap([{ task_id: 'other-task', total_cost_usd: 0.99 }]),
+    });
+    expect(screen.queryByTestId('cost-badge')).not.toBeInTheDocument();
+  });
+
+  it('hides the cost badge when cost is zero', () => {
+    renderCard({}, {
+      costMap: makeCostMap([{ task_id: 'task-1', total_cost_usd: 0 }]),
+    });
+    expect(screen.queryByTestId('cost-badge')).not.toBeInTheDocument();
+  });
+
+  it('hides the cost badge when costMap is undefined', () => {
+    renderCard();
+    expect(screen.queryByTestId('cost-badge')).not.toBeInTheDocument();
+  });
+
+  it('formats cost above one dollar with two decimals', () => {
+    renderCard({}, {
+      costMap: makeCostMap([{ task_id: 'task-1', total_cost_usd: 12.5 }]),
+    });
+    expect(screen.getByTestId('cost-badge').textContent).toContain('$12.50');
+  });
+
+  it('shows sub-cent costs at four-decimal precision', () => {
+    // Regression: $0.0042 must not collapse to $0.00 just because the
+    // badge is visible. Matches TopTasksTable's precision.
+    renderCard({}, {
+      costMap: makeCostMap([{ task_id: 'task-1', total_cost_usd: 0.0042 }]),
+    });
+    const text = screen.getByTestId('cost-badge').textContent ?? '';
+    expect(text).toContain('$0.0042');
+    expect(text).not.toMatch(/\$0\.00\b/);
   });
 });
