+++ web-ui/src/__tests__/components/workspace/QuickActions.test.tsx
