+++ web-ui/src/__tests__/hooks/useEventSource.test.ts
