+++ web-ui/src/__tests__/lib/eventStyles.test.ts
